requirements.txt
fastapi
ipykernel
pip-chill
python-dotenv
swarm
tiktoken
uvicorn
